TFP-6390 legg om til inntektV2
