+++ src/main/java/no/nav/familie/inntektsmelding/integrasjoner/inntektskomponent/InntektskomponentKlient.java
@@ -1,16 +1,14 @@
 package no.nav.familie.inntektsmelding.integrasjoner.inntektskomponent;
 
+import java.math.BigDecimal;
 import java.time.YearMonth;
-import java.util.Collections;
+import java.util.List;
 
 import jakarta.enterprise.context.ApplicationScoped;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import no.nav.tjenester.aordningen.inntektsinformasjon.Aktoer;
-import no.nav.tjenester.aordningen.inntektsinformasjon.request.HentInntektListeBolkRequest;
-import no.nav.tjenester.aordningen.inntektsinformasjon.response.HentInntektListeBolkResponse;
 import no.nav.vedtak.exception.IntegrasjonException;
 import no.nav.vedtak.felles.integrasjon.rest.RestClient;
 import no.nav.vedtak.felles.integrasjon.rest.RestClientConfig;
@@ -19,9 +17,9 @@ import no.nav.vedtak.felles.integrasjon.rest.RestRequest;
 import no.nav.vedtak.felles.integrasjon.rest.TokenFlow;
 
 @ApplicationScoped
-@RestClientConfig(tokenConfig = TokenFlow.AZUREAD_CC, endpointProperty = "hentinntektlistebolk.url",
-    endpointDefault = "https://ikomp.prod-fss-pub.nais.io/rs/api/v1/hentinntektlistebolk",
-    scopesProperty = "hentinntektlistebolk.scopes", scopesDefault = "api://prod-fss.team-inntekt.ikomp/.default")
+@RestClientConfig(tokenConfig = TokenFlow.AZUREAD_CC, endpointProperty = "inntekt.url",
+    endpointDefault = "https://ikomp.prod-fss-pub.nais.io/rest/v2/inntekt",
+    scopesProperty = "inntekt.scopes", scopesDefault = "api://prod-fss.team-inntekt.ikomp/.default")
 public class InntektskomponentKlient {
     private static final Logger LOG = LoggerFactory.getLogger(InntektskomponentKlient.class);
     private static final YearMonth INNTK_TIDLIGSTE_DATO = YearMonth.of(2015, 7);
@@ -38,30 +36,33 @@ public class InntektskomponentKlient {
         this.restConfig = RestConfig.forClient(this.getClass());
     }
 
-    public HentInntektListeBolkResponse finnInntekt(FinnInntektRequest finnInntektRequest) {
+    public List<Inntektsinformasjon> finnInntekt(FinnInntektRequest finnInntektRequest) {
         var request = lagRequest(finnInntektRequest);
         LOG.info("Henter inntekt");
 
-        HentInntektListeBolkResponse response;
         try {
-            response = restClient.send(request, HentInntektListeBolkResponse.class);
+            return restClient.sendReturnOptional(request, InntektApiUt.class)
+                .map(InntektApiUt::data).orElseGet(List::of);
         } catch (RuntimeException e) {
             throw new IntegrasjonException("FP-824246",
                 "Feil ved kall til inntektstjenesten. Meld til #team_registre og #produksjonshendelser hvis dette skjer over lengre tidsperiode.", e);
         }
-        return response;
     }
 
     private RestRequest lagRequest(FinnInntektRequest finnInntektRequest) {
-        var request = new HentInntektListeBolkRequest();
+        var request = new InntektskomponentKlient.InntektApiInn(finnInntektRequest.aktørId(), "8-28", "Foreldrepenger",
+            finnInntektRequest.fom().isAfter(INNTK_TIDLIGSTE_DATO) ? finnInntektRequest.fom() : INNTK_TIDLIGSTE_DATO,
+            finnInntektRequest.tom().isAfter(INNTK_TIDLIGSTE_DATO) ? finnInntektRequest.tom() : INNTK_TIDLIGSTE_DATO);
 
-        request.setIdentListe(Collections.singletonList(Aktoer.newAktoerId(finnInntektRequest.aktørId())));
-        request.setAinntektsfilter("8-28");
-        request.setFormaal("Foreldrepenger");
-
-        request.setMaanedFom(finnInntektRequest.fom().isAfter(INNTK_TIDLIGSTE_DATO) ? finnInntektRequest.fom() : INNTK_TIDLIGSTE_DATO);
-        request.setMaanedTom(finnInntektRequest.tom().isAfter(INNTK_TIDLIGSTE_DATO) ? finnInntektRequest.tom() : INNTK_TIDLIGSTE_DATO);
         return RestRequest.newPOSTJson(request, restConfig.endpoint(), restConfig);
     }
+
+    public record InntektApiInn(String personident, String filter, String formaal, YearMonth maanedFom, YearMonth maanedTom) { }
+
+    public record InntektApiUt(List<Inntektsinformasjon> data) { }
+
+    public record Inntektsinformasjon(YearMonth maaned, String underenhet, List<Inntekt> inntektListe) { }
+
+    public record Inntekt(String type, BigDecimal beloep) { }
 
 }
