+++ src/main/resources/application-dev-gcp.properties
@@ -6,8 +6,6 @@ pdl.scopes=api://dev-fss.pdl.pdl-api/.default
 arbeidsgiver.notifikasjon.url=http://notifikasjon-produsent-api.fager/api/graphql
 arbeidsgiver.notifikasjon.scopes=api://dev-gcp.fager.notifikasjon-produsent-api/.default
 
-hentinntektlistebolk.url=https://ikomp-q2.dev-fss-pub.nais.io/rs/api/v1/hentinntektlistebolk
-hentinntektlistebolk.scopes=api://dev-fss.team-inntekt.ikomp-q2/.default
 inntekt.url=https://ikomp-q2.dev-fss-pub.nais.io/rest/v2/inntekt
 inntekt.scopes=api://dev-fss.team-inntekt.ikomp-q2/.default
