+++ src/main/resources/application-prod-gcp.properties
@@ -6,8 +6,6 @@ pdl.scopes=api://prod-fss.pdl.pdl-api/.default
 arbeidsgiver.notifikasjon.url=http://notifikasjon-produsent-api.fager/api/graphql
 arbeidsgiver.notifikasjon.scopes=api://prod-gcp.fager.notifikasjon-produsent-api/.default
 
-hentinntektlistebolk.url=https://ikomp.prod-fss-pub.nais.io/rs/api/v1/hentinntektlistebolk
-hentinntektlistebolk.scopes=api://prod-fss.team-inntekt.ikomp/.default
 inntekt.url=https://ikomp.prod-fss-pub.nais.io/rest/v2/inntekt
 inntekt.scopes=api://prod-fss.team-inntekt.ikomp/.default
