+++ src/test/resources/application-vtp.properties
@@ -17,8 +17,6 @@ pdl.scopes=api://vtp.teamforeldrepenger.vtp/.default
 arbeidsgiver.notifikasjon.url=http://localhost:8060/rest/api/fager/graphql
 arbeidsgiver.notifikasjon.scopes=api://vtp.teamforeldrepenger.vtp/.default
 
-hentinntektlistebolk.url=http://localhost:8060/rest/inntektskomponenten-ws/rs/api/v1/hentinntektlistebolk
-hentinntektlistebolk.scopes=testscope
 inntekt.url=http://localhost:8060/rest/inntektskomponenten/v2/inntekt/bulk
 inntekt.scopes=testscope
